fix sonar finding with cycles in file imports
